requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.10.4
atomicwrites==1.4.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
databases==0.9.0
fastapi==0.115.3
forex-types==0.0.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
magic-kind==0.2.2
more-itertools==10.5.0
numpy==1.26.4
oanda-candles==0.1.0
oandapyV20==0.6.3
packaging==24.1
pandas==2.2.3
pandas_ta==0.3.14b0
pluggy==0.13.1
py==1.11.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==5.4.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
time-int==0.0.9
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
